Reorganize and Make Code Runnable
Previous version of the repository could not actually be run, as the file addresses were inaccurate, certain sections of the code that were necessary were commented out, and the files needed for zip matching were missing. As a first step, I fix these issues and push these changes. Note that the code for Table 10 has not been  tested.