get unique adsprocessed
adsprocessed has a fair amount of duplicates, which we talked about removing. the primary source is based on SEQUENCE.x

SEQUENCE.x indicates the number of visits one potential buyer has made to a house, and goes up sequentially for each visit. Using just the first visit reduces adsprocessed to 5830. There are a few other duplicates

1. 9 instances of HCITY differing within otherwise identical rows
2. 9 instances of HHMTYPE differing within otherwise identical rows
3. 10 rows that are complete duplicates of another row (likely differed on some value that was not controlled for and thus discarded)/

I used the original adsprocessed_JPE with has PII in it and thus have not reuploaded it.